Add active subscriptions intercom property
